Add all minor changes to prep this branch for merge with main before refactor core type definitions
