+++ src/components/ShowGameData/ShowGameData.tsx
@@ -2,7 +2,7 @@ import React, {useState} from "react";
 import {IShowGameInfoProps} from "../../types/Main.types";
 import * as S from "./ShowGameData.Styles"
 import {IMove} from "../../lib/models/game";
-import CreateMatrix from "../moveMatrix/createMatrix";
+import CreateMatrix from "../CreateMoveMatrix/createMatrix";
 
 
 const ShowGameData: React.FC<IShowGameInfoProps> = ({gameData}) => {
@@ -46,8 +46,7 @@ const ShowGameData: React.FC<IShowGameInfoProps> = ({gameData}) => {
     }
     return (
         <S.MainContainer>
-            Handle showing multiple uploads later
-            <ShowMoves />
+            {/*<ShowMoves />*/}
 
             <S.MenuTabCont>
                 <button
